I removed the pokemon API fetching from the MainActivity class to a seperate class to make MainAcitivity more readable.
